Create stand alone component for GoogleSignInButton
